+++ packages/diagram/src/LikeC4CustomColors.tsx
@@ -1,4 +1,5 @@
 import { type ThemeColorValues } from '@likec4/core'
+import { useMantineStyleNonce } from '@mantine/core'
 import { entries } from 'remeda'
 import type { LikeC4CustomColorsProperties } from './LikeC4CustomColors.props'
 import { vars } from './theme-vars'
@@ -29,9 +30,11 @@ export function LikeC4CustomColors({ customColors }: LikeC4CustomColorsPropertie
     .map(([name, color]) => toStyle(name, color))
     .join('\n')
 
+  const nonce = useMantineStyleNonce()?.()
+
   return (
     <>
-      <style type="text/css" dangerouslySetInnerHTML={{ __html: styles }} />
+      <style type="text/css" dangerouslySetInnerHTML={{ __html: styles }} {...(nonce && { nonce })} />
     </>
   )
 }