+++ packages/likec4/src/vite/config-app.ts
@@ -81,9 +81,12 @@ export const viteConfig = async ({ languageServices, likec4AssetsDir, ...cfg }: 
         'react',
         'react-dom',
         'react/jsx-runtime',
-        'react-dom/client'
-      ],
-      force: true
+        'react/jsx-dev-runtime',
+        'react-dom/client',
+        '@mantine/core',
+        '@mantine/hooks',
+        '@tabler/icons-react'
+      ]
     },
     base,
     build: {